pyproject.toml
[project]
name = "quizbot"
version = "0.1.0"
description = "Create quizzes in Telegram"
authors = [{ name = "Antony Kamp", email = "antony.kamp@student.hpi.de" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[webhooks]~=21.10",
    "sqlalchemy~=2.0",
    "psycopg2-binary~=2.9",
    "alembic~=1.15",
    "python-dotenv~=1.1",
]

[project.urls]
Homepage = "https://github.com/antonykamp/QuizBot"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
